Add init structure to support injecting markdownDescription dynamically
